+++ react/utils/messages.ts
@@ -150,4 +150,5 @@ export const messages = defineMessages({
   orderStatusPaidLabel: { id: 'admin/order.status.paid.label' },
   orderStatusCancelLabel: { id: 'admin/order.status.cancel.label' },
   orderStatusInvoicedLabel: { id: 'admin/order.status.invoiced.label' },
+  totalOrdersLabel: { id: 'admin/total.orders.label' },
 })