+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Totalizers area for the affiliate orders table
+- Totalizers variable to affiliatesOrders query
+
 ## [0.43.0] - 2022-03-10
 
 ### Added